ref[components][structure][linter]: use atomics folder structure & add default import path & adjust linter & run prettier on save
