+++ .eslintignore
@@ -1,2 +1,5 @@
 node_modules
 out
+.husky
+.vscode
+public
