+++ .eslintrc
@@ -1,71 +1,78 @@
 {
-  // Configuration for JavaScript files
-  "extends": [
-    "airbnb-base",
-    "next/core-web-vitals", // Needed to avoid warning in next.js build: 'The Next.js plugin was not detected in your ESLint configuration'
-    "plugin:prettier/recommended"
-  ],
+  "extends": ["airbnb-base", "next/core-web-vitals", "plugin:prettier/recommended"],
   "rules": {
     "prettier/prettier": [
       "error",
       {
+        "printWidth": 120,
         "singleQuote": true,
-        "endOfLine": "auto"
-      }
-    ]
+        "trailingComma": "all",
+        "arrowParens": "avoid",
+        "semi": false,
+        "endOfLine": "auto",
+      },
+    ],
   },
   "overrides": [
-    // Configuration for TypeScript files
     {
       "files": ["**/*.ts", "**/*.tsx"],
-      "plugins": [
-        "@typescript-eslint",
-        "unused-imports",
-        "tailwindcss",
-        "simple-import-sort"
-      ],
+      "plugins": ["@typescript-eslint", "unused-imports", "tailwindcss", "simple-import-sort"],
       "extends": [
         "plugin:tailwindcss/recommended",
         "airbnb-typescript",
         "next/core-web-vitals",
-        "plugin:prettier/recommended"
+        "plugin:prettier/recommended",
       ],
       "parserOptions": {
-        "project": "./tsconfig.json"
+        "project": "./tsconfig.json",
       },
       "rules": {
-        "prettier/prettier": [
+        "import/extensions": "off",
+        "react/destructuring-assignment": "off",
+        "react/require-default-props": "off",
+        "react/jsx-props-no-spreading": "off",
+        "react-hooks/exhaustive-deps": "off",
+        "@next/next/no-img-element": "off",
+        "@next/next/link-passhref": "off",
+        "@typescript-eslint/comma-dangle": "off",
+        "comma-dangle": "off",
+        "@typescript-eslint/consistent-type-imports": "error",
+        "no-restricted-syntax": ["error", "ForInStatement", "LabeledStatement", "WithStatement"],
+        "import/prefer-default-export": "off",
+        "tailwindcss/no-custom-classname": "off",
+        "simple-import-sort/imports": [
           "error",
           {
-            "singleQuote": true,
-            "endOfLine": "auto"
-          }
+            "groups": [
+              // Packages `react` related packages come first.
+              ["^react", "^\\w", "^@hookform", "^@radix-ui"],
+              // npm packages
+              // Anything that starts with a letter (or digit or underscore), or `@` followed by a letter.
+              // ['^\\w'],
+              // Internal packages.
+              ["^@/hooks(/.*|$)"],
+              ["^@/utils(/.*|$)"],
+              ["^@/libs(/.*|$)"],
+              // Internal ui components
+              ["^@/(components|atoms|molecules|globals|organisms|templates|pages)(/.*|$)"],
+
+              // Side effect imports.
+              ["^\\u0000"],
+              // Parent imports. Put `..` last.
+              ["^\\.\\.(?!/?$)", "^\\.\\./?$"],
+              // Other relative imports. Put same-folder imports and `.` last.
+              ["^\\./(?=.*/)(?!/?$)", "^\\.(?!/?$)", "^\\./?$"],
+              // Style imports.
+              ["^.+\\.?(css)$"],
+            ],
+          },
         ],
-        "react/destructuring-assignment": "off", // Vscode doesn't support automatically destructuring, it's a pain to add a new variable
-        "react/require-default-props": "off", // Allow non-defined react props as undefined
-        "react/jsx-props-no-spreading": "off", // _app.tsx uses spread operator and also, react-hook-form
-        "react-hooks/exhaustive-deps": "off", // Incorrectly report needed dependency with Next.js router
-        "@next/next/no-img-element": "off", // We currently not using next/image because it isn't supported with SSG mode
-        "@next/next/link-passhref": "off", // Only needed when the child of Link wraps an <a> tag
-        "@typescript-eslint/comma-dangle": "off", // Avoid conflict rule between Eslint and Prettier
-        "@typescript-eslint/consistent-type-imports": "error", // Ensure `import type` is used when it's necessary
-        "no-restricted-syntax": [
-          "error",
-          "ForInStatement",
-          "LabeledStatement",
-          "WithStatement"
-        ], // Overrides Airbnb configuration and enable no-restricted-syntax
-        "import/prefer-default-export": "off", // Named export is easier to refactor automatically
-        "tailwindcss/no-custom-classname": "off", // Disabled otherwise nightmare to allow each custom tailwind classes
-        "simple-import-sort/imports": "error", // Import configuration for `eslint-plugin-simple-import-sort`
-        "simple-import-sort/exports": "error", // Export configuration for `eslint-plugin-simple-import-sort`
+        "simple-import-sort/exports": "error",
         "@typescript-eslint/no-unused-vars": "off",
         "unused-imports/no-unused-imports": "error",
-        "unused-imports/no-unused-vars": [
-          "error",
-          { "argsIgnorePattern": "^_" }
-        ]
-      }
-    }
-  ]
+        "unused-imports/no-unused-vars": ["error", { "argsIgnorePattern": "^_" }],
+        "semi": "off", // Explicitly disable semicolon enforcement in TypeScript files
+      },
+    },
+  ],
 }
