+++ .vscode/settings.json
@@ -4,24 +4,30 @@
   "search.exclude": {
     "package-lock.json": true
   },
-  "editor.defaultFormatter": "dbaeumer.vscode-eslint",
-  "editor.formatOnSave": false,
-  "editor.codeActionsOnSave": [
-    "source.addMissingImports",
-    "source.fixAll.eslint"
-  ],
+  "eslint.validate": ["javascript", "javascriptreact", "typescript", "typescriptreact"],
+  "[javascript]": {
+    "editor.formatOnSave": true
+  },
+  "[typescript]": {
+    "editor.formatOnSave": true
+  },
+  "editor.defaultFormatter": "esbenp.prettier-vscode", // Set Prettier as the default formatter for all file types
+  "editor.formatOnSave": true, // Enable format on save globally
+  "editor.codeActionsOnSave": {
+    "source.addMissingImports": true, // Continue to add missing imports
+    "source.fixAll.eslint": true // Apply ESLint fixes on save
+  },
   "typescript.tsdk": "node_modules/typescript/lib", // Use the workspace version of TypeScript
-  "typescript.enablePromptUseWorkspaceTsdk": true, // For security reasons it's require that users opt into using the workspace version of typescript
+  "typescript.enablePromptUseWorkspaceTsdk": true, // Prompt to use the workspace version of TypeScript for security
   "jest.autoRun": {
     "watch": false // Start the jest with the watch flag
-    // "onStartup": ["all-tests"] // Run all tests upon project launch
+    // "onStartup": ["all-tests"] // Option to run all tests upon project launch
   },
   "jest.showCoverageOnLoad": true, // Show code coverage when the project is launched
   "jest.autoRevealOutput": "on-exec-error", // Don't automatically open test explorer terminal on launch
-  // Multiple language settings for json and jsonc files
   "[json][jsonc][yaml]": {
-    "editor.formatOnSave": true,
-    "editor.defaultFormatter": "esbenp.prettier-vscode"
+    "editor.formatOnSave": true, // Ensure JSON, JSONC, and YAML files are formatted on save
+    "editor.defaultFormatter": "esbenp.prettier-vscode" // Use Prettier for these file types as well
   },
   "prettier.ignorePath": ".gitignore" // Don't run prettier for files listed in .gitignore
 }
