+++ lint-staged.config.js
@@ -2,4 +2,4 @@ module.exports = {
   '*.{js,jsx,ts,tsx}': ['eslint --fix', 'eslint'],
   '**/*.ts?(x)': () => 'npm run check-types',
   '*.json': ['prettier --write'],
-};
+}
