+++ src/components/globals/background/Background.tsx
@@ -1,0 +1,10 @@
+import type { ReactNode } from 'react'
+
+type IBackgroundProps = {
+  children: ReactNode
+  color: string
+}
+
+const Background = (props: IBackgroundProps) => <div className={props.color}>{props.children}</div>
+
+export { Background }
