+++ src/components/globals/index.ts
@@ -1,0 +1,3 @@
+export * from './background/Background'
+export * from './layout/Meta'
+export * from './layout/Section'
