+++ src/components/organisms/index.ts
@@ -1,0 +1,6 @@
+export * from './feature/VerticalFeatureRow'
+export * from './footer/CenteredFooter'
+export * from './footer/FooterCopyright'
+export * from './footer/FooterIconList'
+export * from './hero/HeroOneButton'
+export * from './navigation/NavbarTwoColumns'
