+++ src/components/templates/Sponsors.tsx
@@ -1,0 +1,9 @@
+import { Section } from '@/components/globals/layout/Section'
+
+const Sponsors = () => (
+  <Section title="Coming soon">
+    <div>Form Coming Soon</div>
+  </Section>
+)
+
+export { Sponsors }
