+++ src/components/templates/VerticalFeatures.tsx
@@ -1,0 +1,40 @@
+import Link from 'next/link'
+
+import { Button } from '@/components/atoms/button/Button'
+import { CTABanner } from '@/components/molecules/cta/CTABanner'
+import { Section } from '@/globals/layout/Section'
+
+const VerticalFeatures = () => (
+  <Section title="" description="">
+    <CTABanner
+      title="More information coming soon"
+      subtitle="Please register for more information"
+      button={
+        <Link href="https://creativedesignsguru.com/category/nextjs/">
+          <Button>Get Started</Button>
+        </Link>
+      }
+    />
+    {/* <VerticalFeatureRow
+      title="Your title here"
+      description="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse bibendum, nunc non posuere consectetur, justo erat semper enim, non hendrerit dui odio id enim."
+      image="/assets/images/feature.svg"
+      imageAlt="First feature alt text"
+    />
+    <VerticalFeatureRow
+      title="Your title here"
+      description="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse bibendum, nunc non posuere consectetur, justo erat semper enim, non hendrerit dui odio id enim."
+      image="/assets/images/feature2.svg"
+      imageAlt="Second feature alt text"
+      reverse
+    />
+    <VerticalFeatureRow
+      title="Your title here"
+      description="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse bibendum, nunc non posuere consectetur, justo erat semper enim, non hendrerit dui odio id enim."
+      image="/assets/images/feature3.svg"
+      imageAlt="Third feature alt text"
+    /> */}
+  </Section>
+)
+
+export { VerticalFeatures }
