+++ src/components/templates/index.ts
@@ -1,0 +1,7 @@
+export * from './Banner'
+export * from './Base'
+export * from './Footer'
+export * from './Hero'
+export * from './Logo'
+export * from './MailChimp'
+export * from './VerticalFeatures'
