+++ src/styles/global.css
@@ -1,5 +1,3 @@
 @tailwind base;
-
 @tailwind components;
-
 @tailwind utilities;
