--- src/templates/Hero.tsx
@@ -1,34 +1,0 @@
-import Link from 'next/link';
-
-import { Background } from '../background/Background';
-import { HeroOneButton } from '../hero/HeroOneButton';
-import { Section } from '../layout/Section';
-import { NavbarTwoColumns } from '../navigation/NavbarTwoColumns';
-import { Logo } from './Logo';
-
-const Hero = () => (
-  <Background color="bg-gray-100">
-    <Section yPadding="py-6">
-      <NavbarTwoColumns logo={<Logo xl />}>
-        <li>
-          <Link href="https://www.rga.com/">Digital Natives</Link>
-        </li>
-      </NavbarTwoColumns>
-    </Section>
-
-    <Section yPadding="pt-20 pb-32">
-      <HeroOneButton
-        title={
-          <>
-            {'A boilerplate for\n'}
-            <span className="text-primary-500">Digital Natives</span>
-          </>
-        }
-        description=""
-        button={''}
-      />
-    </Section>
-  </Background>
-);
-
-export { Hero };
