--- src/templates/Sponsors.tsx
@@ -1,9 +1,0 @@
-import { Section } from '@/layout/Section';
-
-const Sponsors = () => (
-  <Section title="Coming soon">
-    <div>Form Coming Soon</div>
-  </Section>
-);
-
-export { Sponsors };
