+++ src/utils/AppConfig.ts
@@ -1,6 +1,7 @@
 export const AppConfig = {
-  site_name: 'Starter',
-  title: 'Nextjs Starter',
-  description: 'Starter code for your Nextjs Boilerplate with Tailwind CSS',
+  site_name: 'Digital Native',
+  title: 'Digital Native',
+  // TODO: add the correct description
+  description: '...digital native descriptions goes here',
   locale: 'en',
-};
+}
